feat: Add TraceAnything, a new generic tracing engine
TraceAnything does what it says on the tin.  You can feed it a class,
an object, a member, a prototype, on an HTML element type, and it will
explore the properties, methods, and events of the thing, shim them,
and trace them.  It is configurable, and has support for
non-enumerable events and arbitrary logging mechanisms via that
configuration.

Change-Id: Ie90364f4d6287662910523992c6e9318bb548c42