feat: Replace the tracing and log window implementations
Rather than explicitly map out the API of EME and its various
structures, configure TraceAnything with a few different options for
various parts of the API.  The generic tracing engine of TraceAnything
will map out the API and any future additions to these classes
automatically.  New classes or top-level entrypoint can easily be
added to the config in the future.

The new log window implementation is better organized and written as
an ES6 class, and matches the log changes from the tracing part of the
extension.

Change-Id: Iaf5e0a5da7f4500fad14487f8afe27b6610e8634