feat: Instance ID tracking
Each tracked class can have an associated ID field, which defaults to
"id".  If the ID field is null or doesn't exist on the instance, a
generated ID will be assigned instead.

For a class called "Foo", the generated IDs would be of the form
"Foo_1".  Each class will have its own counters that start at 1 for
the first instance logged.  These IDs should be easy to search for in
the log window to correlate calls and events involving specific
instances.

The generated IDs will be shown in the log window in two ways:
 1. On a call or event, for the instance the log is tied to
 2. In serialized data for a call or event (arguments, return values,
    etc)

Change-Id: I189059ffda90adf8385b50325333750476cdf8f6