feat: New formatter API
This revamps the formatter API so that we can receive rich objects
from formatter extensions instead of just strings.  This then allows
us to format things much more nicely in the log window.

This also adds methods to the formatter API to format calls to
MediaKeys.setServerCertficate() and MediaKeySession.generateRequest().

Backward compatibility with the v2 API is preserved, but its usage is
now discouraged, and all documentation has been updated to refer to
this new v3.1 API.

Change-Id: I1a0a9188f1d912f6581394906383e24bdc8edbde