refactor: Restructure service worker and log window (#53)
Originally, the "background page" was responsible for writing files
directly to disk.  That meant sending logs from the log window to the
background page.

Even before moving to manifest v3 and changing the "background page"
to a Service Worker, we stopped writing directly to disk.

After moving to manifest v3, we were sending log messages from the
content script to the log window, then adding them to the DOM and
sending a text version of the log to the service worker.  On download,
the log window would ask the service worker to send the text logs back
to be downloaded by the log window to a file.  This didn't make much
sense.

It was also confusing how "log-window.js" was the service worker, and
"log.js" was the log window.

This renames the service worker to "service-worker.js", and the log
window to "log-window.js".  Now the service worker is only responsible
for starting the log window, and nothing else.  The tests have also
been split up accordingly.

The diff is terrible because of the renaming, but the code is cleaner
and easier to understand IMO.